🚚 move [#2] : 서비스의 연결된 DTO 클래스 위치 변경 - application의 서비스와 밀접하게 연결된 DTO클래스이기 때문에 application 계층으로 이동
