+++ src/main/java/com/sparta/limited/auth_service/application/service/AuthService.java
@@ -1,12 +1,12 @@
 package com.sparta.limited.auth_service.application.service;
 
+import com.sparta.limited.auth_service.application.dto.request.AuthSignupRequest;
+import com.sparta.limited.auth_service.application.dto.response.AuthSignupResponse;
 import com.sparta.limited.auth_service.infrastructure.dto.request.UserCreateInternalRequest;
 import com.sparta.limited.auth_service.infrastructure.dto.response.UserCreateInternalResponse;
 import com.sparta.limited.auth_service.infrastructure.feign.AuthDtoMapper;
 import com.sparta.limited.auth_service.infrastructure.feign.UserFeignClient;
 import com.sparta.limited.auth_service.infrastructure.password.PasswordEncoderUtil;
-import com.sparta.limited.auth_service.presentation.dto.request.AuthSignupRequest;
-import com.sparta.limited.auth_service.presentation.dto.response.AuthSignupResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
